refactor slideviewer download
